+++ src/Events/UserRegistered.php
@@ -11,6 +11,6 @@ class UserRegistered
 
     public function __construct(
         public User $user,
-        public ?int $tenantId = null
+        public ?string $tenantId = null
     ) {}
 }