+++ src/Jobs/SendPasswordResetJob.php
@@ -23,7 +23,7 @@ class SendPasswordResetJob implements ShouldQueue
     use Dispatchable, InteractsWithQueue, Queueable, SerializesModels;
 
     public int $tries = 3;
-    public int $backoff = 30;
+    public array $backoff = [10, 30, 60];
 
     public function __construct(
         public int $userId